+++ backend/learn-dto/src/main/java/com/prosper/learn/dto/CourseDTO.java
@@ -5,17 +5,17 @@ import lombok.Data;
 @Data
 public class CourseDTO {
 
-    private Integer id;
+    private Long id;
 
     private String name;
 
     private String description;
 
-    private Integer creator;
+    private Long creator;
 
-    private Integer rootNode;
+    private Long rootNode;
 
-    private Integer parentId;
+    private Long parentId;
 
     private String state; // 修改为 String 类型，支持 SUMMITTED, APPROVED, REJECTED
