+++ backend/learn-dto/src/main/java/com/prosper/learn/dto/message/CommentMessageDTO.java
@@ -11,6 +11,6 @@ public class CommentMessageDTO extends MessageDTO {
 
     UserDTOV4 commenter;
 
-    int commentId;
+    Long commentId;
 
 }
