+++ backend/learn-dto/src/main/java/com/prosper/learn/dto/message/UpvoteMessageDTO.java
@@ -9,11 +9,11 @@ public class UpvoteMessageDTO extends MessageDTO {
 
     NodeDTOV1 node;
 
-    int objectId;
+    Long objectId;
 
-    int objectType;
+    Integer objectType;
 
-    int voteType;
+    Integer voteType;
 
     UserDTOV4 upvoter;
 }
